data-management-service/requirements.txt
Flask==2.0.1
google-cloud-storage==1.42.0
pandas==1.3.3
gunicorn==20.1.0
